receivables/requirements.txt
functions-framework==3.*
google-cloud-storage
bs4
google-cloud-firestore